requirements.txt
Flask==3.0.3
boto3==1.34.131
psycopg2-binary==2.9.9
python-dotenv==1.0.1
